question_generator/requirements.txt
absl-py==0.13.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
argcomplete==1.12.3
argon2-cffi==20.1.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==19.3.0
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.0.0
blis==0.4.1
boto==2.49.0
boto3==1.18.16
botocore==1.21.16
cached-property==1.5.2
cachetools==4.2.2
catalogue==1.0.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cloudpickle==0.6.1
coloredlogs==10.0
constantly==15.1.0
crcmod==1.7
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.6
dataclasses==0.8
datasets==1.16.1
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
docformatter==1.4
docutils==0.15.2
emoji==1.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
fasteners==0.16.3
filelock==3.0.12
flake8==3.9.2
flake8-bugbear==21.4.3
frozenlist==1.2.0
fsspec==2022.1.0
future==0.17.1
fuzzywuzzy==0.18.0
fvcore==0.1.5.post20210804
gast==0.5.2
gcs-oauth2-boto-plugin==2.7
gdown==3.13.0
gevent==1.5.0
gitdb==4.0.7
gitdb2==4.0.2
GitPython==3.1.18
google-apitools==0.5.32
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
google-reauth==0.1.1
greenlet==1.1.1
grpcio==1.39.0
gsutil==4.66
h5py==3.1.0
httplib2==0.19.1
huggingface-hub==0.4.0
humanfriendly==9.2
hyperlink==21.0.0
idna==3.2
idna-ssl==1.1.0
imagesize==1.2.0
importlib-metadata==4.6.3
incremental==21.3.0
iniconfig==1.1.1
iopath==0.1.9
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonlines==2.0.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
klein==17.10.0
Markdown==3.3.4
markdown-it-py==1.1.0
MarkupSafe==2.0.1
matplotlib==2.2.5
mccabe==0.6.1
mdit-py-plugins==0.2.8
mdurl==0.1.0
mistune==0.8.4
mock==2.0.0
monotonic==1.6
multidict==5.2.0
multiprocess==0.70.12.2
murmurhash==1.0.6
myst-parser==0.15.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.3
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
packaging==18.0
pandas==1.1.5
pandocfilters==1.4.3
parlai==0.9.4
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plac==1.1.3
pluggy==0.13.1
portalocker==2.3.0
preshed==3.0.6
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
py==1.10.0
py-gfm==1.0.2
py-rouge==1.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.2.4
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytorch-lightning==0.7.5
pytorch-transformers==1.1.0
pytz==2021.1
pyu2f==0.1.5
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.9.0
rasa-nlu==0.15.1
regex==2021.8.3
requests==2.26.0
requests-mock==1.9.3
requests-oauthlib==1.3.0
retry-decorator==1.1.1
rsa==4.7.2
ruamel.yaml==0.15.100
s3transfer==0.5.0
sacremoses==0.0.47
scikit-learn==0.20.4
scipy==1.5.4
semantic-text-similarity==1.0.3
Send2Trash==1.8.0
sentencepiece==0.1.96
sh==1.14.2
simplejson==3.17.3
six==1.16.0
sklearn-crfsuite==0.3.6
smmap==4.0.0
snowballstemmer==2.1.0
spacy==2.3.1
Sphinx==4.1.2
sphinx-autodoc-typehints==1.12.0
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==1.0.5
strsim==0.0.3
styleformer @ git+https://github.com/PrithivirajDamodaran/Styleformer.git@505ca67741d4f4e4e6f1997e3842b30e3f9a21f8
subword-nmt==0.3.7
tabulate==0.8.9
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
thinc==7.4.1
threadpoolctl==2.2.0
tokenizers==0.8.0rc4
toml==0.10.2
tomli==1.2.3
torch==1.10.2
torchtext==0.11.2
tornado==6.1
tqdm==4.62.0
traitlets==4.3.3
transformers==3.0.0
Twisted==21.7.0
typing==3.7.4.3
typing-extensions==3.10.0.0
Unidecode==1.2.0
untokenize==0.1.1
urllib3==1.26.6
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.1
websocket-server==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xxhash==2.0.2
yacs==0.1.8
yarl==1.7.2
zipp==3.5.0
zope.interface==5.4.0
